Split analysis claim error logging
